scripts: support centos builds

+++ script/cleanup.sh
@@ -29,44 +29,51 @@ echo "==> Cleaning up tmp"
 rm -rf /tmp/*
 
 # Remove some packages to get a minimal install
-echo "==> Removing all linux kernels except the current one"
-dpkg --list | awk '{ print $2 }' | grep 'linux-image-3.*-generic' | grep -v $(uname -r) | xargs apt-get -y purge
-echo "==> Removing linux source"
-dpkg --list | awk '{ print $2 }' | grep linux-source | xargs apt-get -y purge
-echo "==> Removing development packages"
-dpkg --list | awk '{ print $2 }' | grep -- '-dev$' | xargs apt-get -y purge
-echo "==> Removing documentation"
-dpkg --list | awk '{ print $2 }' | grep -- '-doc$' | xargs apt-get -y purge
-echo "==> Removing development tools"
-apt-get -y purge build-essential
-echo "==> Removing default system Ruby"
-apt-get -y purge ruby ri doc
-echo "==> Removing default system Python"
-apt-get -y purge python-dbus libnl1 python-smartpm python-twisted-core libiw30 python-twisted-bin libdbus-glib-1-2 python-pexpect python-pycurl python-serial python-gobject python-pam python-openssl libffi5
-echo "==> Removing X11 libraries"
-apt-get -y purge libx11-data xauth libxmuu1 libxcb1 libx11-6 libxext6
-echo "==> Removing obsolete networking components"
-apt-get -y purge ppp pppconfig pppoeconf
-echo "==> Removing other oddities"
-apt-get -y purge popularity-contest installation-report wireless-tools wpasupplicant
-echo "==> Removing groff info lintian linda"
-rm -rf /usr/share/groff/* /usr/share/info/* /usr/share/lintian/* /usr/share/linda/*
+if [[ -f /etc/centos-release ]]; then
+	echo "==> Purging yum package caches"
+	yum -y clean all
 
-echo "==> Removing man pages"
-find /usr/share/man -type f -delete
-echo "==> Removing APT files"
-find /var/lib/apt -type f -exec rm -f {} +
-echo "==> Removing anything in /usr/src, except linux headers"
-find /usr/src -mindepth 1 -maxdepth 1 -not -name 'linux-headers-*' -exec rm -rf {} +
-echo "==> Removing any docs"
-find /usr/share/doc -type f -delete
-echo "==> Removing caches"
-find /var/cache -type f -exec rm -f {} +
+else
+	echo "==> Removing all linux kernels except the current one"
+	dpkg --list | awk '{ print $2 }' | grep 'linux-image-3.*-generic' | grep -v $(uname -r) | xargs apt-get -y purge
+	echo "==> Removing linux source"
+	dpkg --list | awk '{ print $2 }' | grep linux-source | xargs apt-get -y purge
+	echo "==> Removing development packages"
+	dpkg --list | awk '{ print $2 }' | grep -- '-dev$' | xargs apt-get -y purge
+	echo "==> Removing documentation"
+	dpkg --list | awk '{ print $2 }' | grep -- '-doc$' | xargs apt-get -y purge
+	echo "==> Removing development tools"
+	apt-get -y purge build-essential
+	echo "==> Removing default system Ruby"
+	apt-get -y purge ruby ri doc
+	echo "==> Removing default system Python"
+	apt-get -y purge python-dbus libnl1 python-smartpm python-twisted-core libiw30 python-twisted-bin libdbus-glib-1-2 python-pexpect python-pycurl python-serial python-gobject python-pam python-openssl libffi5
+	echo "==> Removing X11 libraries"
+	apt-get -y purge libx11-data xauth libxmuu1 libxcb1 libx11-6 libxext6
+	echo "==> Removing obsolete networking components"
+	apt-get -y purge ppp pppconfig pppoeconf
+	echo "==> Removing other oddities"
+	apt-get -y purge popularity-contest installation-report wireless-tools wpasupplicant
+	echo "==> Removing groff info lintian linda"
+	rm -rf /usr/share/groff/* /usr/share/info/* /usr/share/lintian/* /usr/share/linda/*
 
-# Cleanup apt cache
-apt-get -y autoremove --purge
-apt-get -y clean
-apt-get -y autoclean
+	echo "==> Removing man pages"
+	find /usr/share/man -type f -delete
+	echo "==> Removing APT files"
+	find /var/lib/apt -type f -exec rm -f {} +
+	echo "==> Removing anything in /usr/src, except linux headers"
+	find /usr/src -mindepth 1 -maxdepth 1 -not -name 'linux-headers-*' -exec rm -rf {} +
+	echo "==> Removing any docs"
+	find /usr/share/doc -type f -delete
+	echo "==> Removing caches"
+	find /var/cache -type f -exec rm -f {} +
+
+	# Cleanup apt cache
+	echo "==> Purging apt package caches"
+	apt-get -y autoremove --purge
+	apt-get -y clean
+	apt-get -y autoclean
+fi
 
 # Remove Bash history
 unset HISTFILE
